requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pymongo>=3.12.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
fpdf2>=2.7.0
